+++ src/test/validator.test.js
@@ -47,6 +47,18 @@ describe('validator', () => {
             '#main-container > div[\\:class^="$test.ad.RenderedDesktop"]',
             'div[class\\"ads-article\\"]',
             "[class\\'ads-article\\']",
+            // TODO: delete this comment later.
+            // considered as invalid by nwsapi until the bugs are fixed
+            // but they should be valid
+            // https://github.com/dperini/nwsapi/issues/55:
+            'div:not(div > span)',
+            'div:not(h1 + p)',
+            // TODO: delete this comment later.
+            // https://github.com/dperini/nwsapi/issues/71:
+            '.i_con + :not(.i_con):not([href="javascript:void(0);"])',
+            '.share--content button:not([onclick="window.print();"])',
+            '.sns > a[href^="javascript:openSendNews"]:not([href="javascript:openSendNews(\'url\');"])',
+            '.social > nav > ul > li > a[href="javascript:;"][onclick^="Share"]:not([onclick="Share(\'P\');"])',
         ];
         test.each(validSelectors)('%s', (selector) => {
             const rules = [`example.com##${selector}`];